requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.7
mypy==1.10.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
pika==1.3.2
pydantic==2.7.4
pydantic_core==2.18.4
python-decouple==3.8
requests==2.32.3
ruff==0.4.10
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
